requirements.txt
Flask
requests
python-dotenv
Flask-SQLAlchemy
flask-login
psycopg2-binary